+++ tests/fixtures/ark.ts
@@ -8,6 +8,8 @@ export const ArkAddresses = {
     },
     valid: {
       bark: 'tark1pem36wcfzqqp546js6p0sfchhf78vadpesshk2w40xglj2y84ywmznsfu7zz8q6fzqypgpq22n7xxy7pu9jachh8m76ua57khufcvltmj8emtpeveh2x7lwqgenwkv',
+      barkUppercase:
+        'TARK1PEM36WCFZQQP546JS6P0SFCHHF78VADPESSHK2W40XGLJ2Y84YWMZNSFU7ZZ8Q6FZQYPGPQ22N7XXY7PU9JACHH8M76UA57KHUFCVLTMJ8EMTPEVEH2X7LWQGENWKV',
       arkade:
         'tark1qplnj2gett9j483fchy6chaxn4y52c4g7n5djh9xua3ywdxw0ldatc3e9xcj9xpx0r5tmr0dgvu2f4s352muklg0tcxx0scnnkraajy9jgz4xl'
     }